Renamed CoreGDBRemote.{h,cpp} to CoreFile.{h,cpp}. Made required changes for this in CMakeLists and in other sources.
