+++ llvm-15.0.3/lldb/tools/lldb-crash-server/CMakeLists.txt
@@ -44,7 +44,7 @@ add_lldb_tool(lldb-crash-server
     lldb-server.cpp
     LLDBServerUtilities.cpp
     SystemInitializerLLGS.cpp
-    CoreGDBRemote.cpp
+    CoreFile.cpp
     GDBRemoteCommunicationServerCS.cpp
 
     LINK_LIBS
