+++ llvm-15.0.3/lldb/tools/lldb-crash-server/CoreFile.h
