+++ llvm-15.0.3/lldb/tools/lldb-crash-server/lldb-gdbserver.cpp
@@ -37,7 +37,7 @@
 #include "llvm/Support/WithColor.h"
 
 #include "GDBRemoteCommunicationServerCS.h"
-#include "CoreGDBRemote.h"
+#include "CoreFile.h"
 
 #if defined(__linux__)
 #include "Plugins/Process/Linux/NativeProcessLinux.h"
